pyproject.toml
[tool.poetry]
name = "hestia-logger"
version = "0.1.0"
description = ""
authors = ["FOX Techniques <ali.nabbi@fox-techniques.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
coloredlogs = "^15.0.1"
elasticsearch = "^8.17.2"
python-json-logger = "^3.2.1"
fastapi = "^0.115.11"
requests = "^2.32.3"
structlog = "^25.1.0"
aiofiles = "^24.1.0"
httpx = "^0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
